mysite/requirement.txt
django==4.0.6
djangorestframework==3.13.1
environs==10.1.1
drf_yasg==1.24.1
django-allauth==0.48.0
corsheaders==4.3.1
